moved shop into its own component
