+++ client/src/components/TestComponent/index.js
@@ -1,0 +1,9 @@
+const TestComponent = () => {
+  return (
+    <>
+      <p>test component rendering</p>
+    </>
+  );
+};
+
+export default TestComponent;
